contributors/adobe/submitted/exclusions/wrap-margin-001.html

<!DOCTYPE HTML>
<html>
    <head>
        <title>CSS Exclusions Test: wrap-margin property set</title>
        <link rel="author" title="Jacob Goldstein" href="mailto:jacobg@adobe.com">
        <link rel="help" href="http://www.w3.org/TR/css-shapes/#shape-margin-property">
        <meta name="flags" content="ahem dom">
        <meta name="assert" content="Text should flow up to the margin defined around the exclusion">
        <script src="/resources/testharness.js"></script>
        <script src="/resources/testharnessreport.js"></script>
        <script src="resources/helper.js"></script>
        <style>
            * {
                margin: 0px;
                padding: 0px;
            }
            .exclusion {
                wrap-flow: start;
                wrap-margin: 12px;
                position: absolute;
                top: 40px;
                left: 40px;
                width: 60px;
                height: 25px;
                background: blue;
            }    
            #content {
                width: 400px;
                line-height: 12px;
                font: 12px Ahem;
            }
        </style>
    </head>
    <body>
        <div class="exclusion"></div>

        <div id="content">
            <span id="linesAbove">ethical<BR>beer<BR>craft<BR></span>
            <span id="lineRight1">foo</span>
            <span id="lineRight2">sed</span>
            <span id="linesBelow">Vegan<BR>cliche<BR>retro</span>
        </div>
        
        <div id="log"></div>
        
        <script type="text/javascript">       
            test(function() {assert_equals(checkLinePos("lineRight1",36,"top"), true)}, "Verify top of the 'lineRight1' span is positioned correctly");
            test(function() {assert_equals(checkLinePos("lineRight1",110,"left"), true)}, "Verify left of the 'lineRight1' span is positioned correctly");
            test(function() {assert_equals(checkLinePos("lineRight2",48,"top"), true)}, "Verify top of the 'lineRight2' span is positioned correctly");
            test(function() {assert_equals(checkLinePos("lineRight2",110,"left"), true)}, "Verify left of the 'lineRight2' span is positioned correctly");
            test(function() {assert_equals(checkLinePos("linesBelow",75,"top"), true)}, "Verify top of the 'lineBelow' span is positioned correctly");
        </script>
    </body>
</html>
